(cleanup) move files for neater structure + rename some files
